api/requirements.txt
Flask
pillow
rembg